xlogparse: initial refactoring of the parsing library
The xlogdump package has a number of routines which come in handy
when you need to parse a postgres WAL journal.  This commit takes
the first step in splitting useful logic out into an xlogparse
library, for use by arbitrary clients, by factoring out the
ReadRecord() routine.  The new routine returns its state rather
than just a pass/fail, allowing clients to customize their
response to conditions like end-of-log, and log file switch.

As part of this refactoring, took the opportunity to cull unused
header inclusions from affected files.
