+++ Makefile.libxlogparse
@@ -1,0 +1,19 @@
+NAME = xlogparse
+OBJS = xlogparse.o
+SO_MAJOR_VERSION = 0
+SO_MINOR_VERSION = 6
+SHLIB_LINK = $(libpq_pgport)
+
+
+all: all-lib
+
+ifdef USE_PGXS
+PGXS := $(shell pg_config --pgxs)
+include $(PGXS)
+else
+subdir = contrib/xlogdump
+top_builddir = ../..
+include $(top_builddir)/src/Makefile.global
+include $(top_srcdir)/contrib/contrib-global.mk
+include $(top_srcdir)/src/Makefile.shlib
+endif
